refactor: split binding source workspace helpers
