+++ packages/wp-typia-project-tools/src/runtime/cli-add-workspace-binding-source-types.ts
@@ -1,0 +1,16 @@
+import type { PostMetaBindingField } from "./post-meta-binding-fields.js";
+
+export type BindingTarget = {
+	attributeName: string;
+	blockSlug: string;
+};
+
+export type BindingPostMetaSource = {
+	fields: PostMetaBindingField[];
+	metaKey: string;
+	metaPath: string;
+	postMetaSlug: string;
+	postType: string;
+	schemaFile: string;
+	sourceTypeName: string;
+};
